requirements.txt
streamlit
openai
google-adk